+++ src/custom/components/Toggle/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import ToggleUni, { ToggleProps as Togg
-lePropsUni } from '@src/components/Toggle'
+import ToggleUni, { ToggleProps as TogglePropsUni } from '@src/components/Toggle'
 
 export type ToggleProps = TogglePropsUni
